+++ tests/error/many_temp_regions_typed.c
@@ -1,4 +1,4 @@
-#include "../../src/koliseo.h"
+#include "../../src/kls_region.h"
 
 int main(void) {